refactor: merge gsheet operation files into sheet.py
Merge clone.py, pull.py, push.py, folder_push.py into sheet.py.
Keep client.py (API wrapper) and frontmatter.py (file format) as
separate stable abstractions.

gsheet/ goes from 7 files to 4:
  __init__.py  (re-exports)
  client.py    (GSheetClient)
  frontmatter.py (SheetConfig)
  sheet.py     (operations + Resource classes)
