+++ gax/gsheet/__init__.py
@@ -4,7 +4,15 @@ Re-exports from submodules. CLI commands live in cli.py.
 """
 
 from .client import GSheetClient  # noqa: F401
-from .pull import pull  # noqa: F401
-from .push import push  # noqa: F401
-from .clone import clone_all, pull_all  # noqa: F401
-from .sheet import SheetTab, Sheet, _extract_spreadsheet_id  # noqa: F401
+from .frontmatter import SheetConfig  # noqa: F401
+from .sheet import (  # noqa: F401
+    SheetTab,
+    Sheet,
+    _extract_spreadsheet_id,
+    clone_all,
+    pull_all,
+    pull_single_tab,
+    pull_single_tab as pull,
+    push_single_tab,
+    push_single_tab as push,
+)
